+++ scripts/constants.js
@@ -52,6 +52,7 @@ const CONSTANTS = {
     HD_MULTIPLIER: "recovery-hitdice",
     LONG_RESOURCES_MULTIPLIER: "recovery-resources",
     LONG_SPELLS_MULTIPLIER: "recovery-spells",
+    LONG_CUSTOM_SPELL_RECOVERY: "long-recovery-custom-spell-points",
     LONG_PACT_SPELLS_MULTIPLIER: "long-recovery-pact-spells",
     LONG_USES_OTHERS_MULTIPLIER: "recovery-uses-others",
     LONG_USES_FEATS_MULTIPLIER: "recovery-uses-feats",
@@ -198,8 +199,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     scope: "world",
     group: "shortrest",
     customSettingsDialog: true,
-    validate: (settingsMap) => {
-      return settingsMap.get(CONSTANTS.SETTINGS.DISABLE_SHORT_REST_HIT_DICE).value;
+    dependsOn: [CONSTANTS.SETTINGS.DISABLE_SHORT_REST_HIT_DICE],
+    validate: (settings) => {
+      return settings.get(CONSTANTS.SETTINGS.DISABLE_SHORT_REST_HIT_DICE).value;
     },
     config: false,
     default: 0,
@@ -211,8 +213,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     scope: "world",
     group: "shortrest",
     customSettingsDialog: true,
-    validate: (settingsMap) => {
-      return settingsMap.get(CONSTANTS.SETTINGS.DISABLE_SHORT_REST_HIT_DICE).value;
+    dependsOn: [CONSTANTS.SETTINGS.DISABLE_SHORT_REST_HIT_DICE],
+    validate: (settings) => {
+      return settings.get(CONSTANTS.SETTINGS.DISABLE_SHORT_REST_HIT_DICE).value;
     },
     customFormula: CONSTANTS.SETTINGS.MAX_HIT_DICE_SPEND_FORMULA,
     config: false,
@@ -224,6 +227,14 @@ CONSTANTS.DEFAULT_SETTINGS = {
       [CONSTANTS.FRACTIONS.CUSTOM]: "REST-RECOVERY.Fractions.Custom",
     },
     default: CONSTANTS.FRACTIONS.FULL,
+  },
+  [CONSTANTS.SETTINGS.MAX_HIT_DICE_SPEND_FORMULA]: {
+    scope: "world",
+    group: "shortrest",
+    config: false,
+    hidden: true,
+    type: String,
+    default: "@details.level",
   },
   [CONSTANTS.SETTINGS.SHORT_RESOURCES_MULTIPLIER]: {
     name: "REST-RECOVERY.Settings.ShortRest.ResourcesRecoveryFraction.Title",
@@ -345,8 +356,8 @@ CONSTANTS.DEFAULT_SETTINGS = {
     hint: "REST-RECOVERY.Settings.LongRest.ExhaustionIntegration.Hint",
     scope: "world",
     group: "longrest",
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.EXHAUSTION_INTEGRATION).value
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.EXHAUSTION_INTEGRATION).value
     },
     config: false,
     type: String,
@@ -372,8 +383,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     hint: "REST-RECOVERY.Settings.LongRest.PreRegainHitDice.Hint",
     scope: "world",
     group: "longrest",
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.LONG_REST_ROLL_HIT_DICE).value
+    dependsOn: [CONSTANTS.SETTINGS.LONG_REST_ROLL_HIT_DICE],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.LONG_REST_ROLL_HIT_DICE).value
     },
     config: false,
     default: false,
@@ -384,8 +396,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     hint: "REST-RECOVERY.Settings.LongRest.PreRegainHitDiceBuffer.Hint",
     scope: "world",
     group: "longrest",
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.PRE_REST_REGAIN_HIT_DICE).value
+    dependsOn: [CONSTANTS.SETTINGS.PRE_REST_REGAIN_HIT_DICE],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.PRE_REST_REGAIN_HIT_DICE).value
     },
     config: false,
     default: false,
@@ -509,6 +522,27 @@ CONSTANTS.DEFAULT_SETTINGS = {
     type: String,
     default: "@slot.max",
   },
+  [CONSTANTS.SETTINGS.LONG_CUSTOM_SPELL_RECOVERY]: {
+    name: "REST-RECOVERY.Settings.LongRest.CustomSpellSlotRecovery.Title",
+    hint: "REST-RECOVERY.Settings.LongRest.CustomSpellSlotRecovery.Hint",
+    dependsOn: [CONSTANTS.SETTINGS.LONG_SPELLS_MULTIPLIER],
+    validate: (settings) => {
+      return settings.get(CONSTANTS.SETTINGS.LONG_SPELLS_MULTIPLIER).value !== CONSTANTS.FRACTIONS.CUSTOM;
+    },
+    callback: (settings) => {
+      const setting = settings.get(CONSTANTS.SETTINGS.LONG_SPELLS_MULTIPLIER_FORMULA);
+      if(setting.value === "@slot.max"){
+        setting.store.set("ceil(min(17, @details.level+1)/2)*2");
+      }
+    },
+    scope: "world",
+    group: "longrest",
+    customSettingsDialog: true,
+    nonDefaultSetting: true,
+    config: false,
+    default: false,
+    type: Boolean
+  },
   [CONSTANTS.SETTINGS.LONG_PACT_SPELLS_MULTIPLIER]: {
     name: "REST-RECOVERY.Settings.LongRest.PactSpellSlotsLongRecoveryFraction.Title",
     hint: "REST-RECOVERY.Settings.LongRest.PactSpellSlotsLongRecoveryFraction.Hint",
@@ -627,8 +661,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
   [CONSTANTS.SETTINGS.LONG_REST_ARMOR_HIT_DICE]: {
     name: "REST-RECOVERY.Settings.LongRest.ArmorHitDiceRecoveryFraction.Title",
     hint: "REST-RECOVERY.Settings.LongRest.ArmorHitDiceRecoveryFraction.Hint",
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.LONG_REST_ARMOR_AUTOMATION).value;
+    dependsOn: [CONSTANTS.SETTINGS.LONG_REST_ARMOR_AUTOMATION],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.LONG_REST_ARMOR_AUTOMATION).value;
     },
     scope: "world",
     group: "longrest",
@@ -657,9 +692,10 @@ CONSTANTS.DEFAULT_SETTINGS = {
   [CONSTANTS.SETTINGS.LONG_REST_ARMOR_EXHAUSTION]: {
     name: "REST-RECOVERY.Settings.LongRest.AutomateArmorExhaustion.Title",
     hint: "REST-RECOVERY.Settings.LongRest.AutomateArmorExhaustion.Hint",
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.AUTOMATE_EXHAUSTION).value
-          || !settingsMap.get(CONSTANTS.SETTINGS.LONG_REST_ARMOR_AUTOMATION).value;
+    dependsOn: [CONSTANTS.SETTINGS.AUTOMATE_EXHAUSTION, CONSTANTS.SETTINGS.LONG_REST_ARMOR_AUTOMATION],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.AUTOMATE_EXHAUSTION).value
+          || !settings.get(CONSTANTS.SETTINGS.LONG_REST_ARMOR_AUTOMATION).value;
     },
     scope: "world",
     group: "longrest",
@@ -814,8 +850,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     scope: "world",
     group: "foodandwater",
     customSettingsDialog: true,
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value
+    dependsOn: [CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value
     },
     config: false,
     default: 1,
@@ -827,8 +864,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     scope: "world",
     group: "foodandwater",
     customSettingsDialog: true,
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value
+    dependsOn: [CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value
     },
     config: false,
     default: 1,
@@ -841,8 +879,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     group: "foodandwater",
     customSettingsDialog: true,
     config: false,
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value
+    dependsOn: [CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value
     },
     choices: {
       [CONSTANTS.FRACTIONS.FULL]: "REST-RECOVERY.Fractions.Full",
@@ -859,8 +898,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     group: "foodandwater",
     customSettingsDialog: true,
     config: false,
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value
+    dependsOn: [CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value
     },
     choices: {
       [CONSTANTS.FRACTIONS.FULL]: "REST-RECOVERY.Fractions.Full",
@@ -875,9 +915,10 @@ CONSTANTS.DEFAULT_SETTINGS = {
     hint: "REST-RECOVERY.Settings.FoodAndWater.AutomateFoodWaterExhaustion.Hint",
     scope: "world",
     group: "foodandwater",
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.AUTOMATE_EXHAUSTION).value
-        || !settingsMap.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value;
+    dependsOn: [CONSTANTS.SETTINGS.AUTOMATE_EXHAUSTION, CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.AUTOMATE_EXHAUSTION).value
+        || !settings.get(CONSTANTS.SETTINGS.ENABLE_FOOD_AND_WATER).value;
     },
     config: false,
     default: false,
@@ -888,8 +929,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     hint: "REST-RECOVERY.Settings.FoodAndWater.NoFoodDuration.Hint",
     scope: "world",
     group: "foodandwater",
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.AUTOMATE_FOODWATER_EXHAUSTION).value
+    dependsOn: [CONSTANTS.SETTINGS.AUTOMATE_FOODWATER_EXHAUSTION],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.AUTOMATE_FOODWATER_EXHAUSTION).value
     },
     config: false,
     default: "3+max(1,@abilities.con.mod)",
@@ -900,8 +942,9 @@ CONSTANTS.DEFAULT_SETTINGS = {
     hint: "REST-RECOVERY.Settings.FoodAndWater.HalfWaterSaveDC.Hint",
     scope: "world",
     group: "foodandwater",
-    validate: (settingsMap) => {
-      return !settingsMap.get(CONSTANTS.SETTINGS.AUTOMATE_FOODWATER_EXHAUSTION).value
+    dependsOn: [CONSTANTS.SETTINGS.AUTOMATE_FOODWATER_EXHAUSTION],
+    validate: (settings) => {
+      return !settings.get(CONSTANTS.SETTINGS.AUTOMATE_FOODWATER_EXHAUSTION).value
     },
     config: false,
     default: 15,